move blog model to doamin layer
