+++ app/src/main/java/com/droid/diexample/data/MainRepository.kt
@@ -1,6 +1,6 @@
 package com.droid.diexample.data
 
-import com.droid.diexample.model.Blog
+import com.droid.diexample.domain.Blog
 import com.droid.diexample.data.remote.api.BlogApiService
 import com.droid.diexample.data.remote.mapper.NetworkMapper
 import com.droid.diexample.data.local.dao.BlogDao
